requirements.txt
Pillow
tqdm
piexif
colorama